Backend-main/requirements.txt
coreapi==2.3.3
datetime-truncate==1.1.1
Django==3.2


gunicorn==21.2.0

django-3-jet==1.0.8
django-cors-headers==4.2.0 
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.7
django-jazzmin==2.6.0

django-ckeditor==6.7.0
django-ratelimit==4.1.0

psycopg2-binary==2.9.7

django-multiselectfield==0.1.12


openai==0.27.8
openapi-codec==1.3.2
passwordgenerator==1.5.1
random-password-generator==2.2.0
Pillow==10.0.0


stripe==5.5.0

django-ratelimit==4.1.0


python-dotenv==1.0.0

google==3.0.0
google-api-core==2.11.1
facebook-sdk==3.1.0
boto3==1.28.51


pytube==15.0.0
moviepy==1.0.3

python-magic==0.4.27
pydub==0.25.1
python-docx==0.8.11

PyPDF2==3.0.1
python-dateutil==2.8.2


sseclient==0.0.27
channels-redis==2.4.2
channels==2.4.0
daphne==2.5.0
decouple==0.0.7
Faker==8.11.0



django-storages==1.14.2